Implement support for sending UDP for macOS
The reason it didn't work is because bigstring needs a way to send
strings in a way where it is not interrupted by SIGPIPE. The proper flag
to do so in GNU/Linux is MSG_NOSIGNAL, but this is not defined on macOS
and probably at least some BSDs. This commit adds support for detecting
the alternate option, SO_NOSIGPIPE and picks that one.

There is another, similar implementation in the Linux_ext module, but
since it is full of GNU/Linux specific code, it was not touched.

Presumably this also works in FreeBSD (which includes SO_NOSIGPIPE) and
DragonFlyBSD (which has both MSG_NOSIGNAL and SO_NOSIGPIPE).