Fix build on MinGW
MinGW does not provide endian.h, similar to MSVC. Changing this ifndef from _MSC_VER to _WIN32 to cover both toolchains.

Test Plan:
I compiled the project (at v0.16 tag) with this change and it compiled, but I have not tested further. I have not attempted to compile with MSVC, but I'm quite confident that _WIN32 will be defined.